python/requirements.txt
cassandra-driver==3.25.0
fastapi==0.78.0
python-dotenv==0.20.0
uvicorn==0.18.2
